재료 테이블 생성하면서 생긴 오류 수정

+++ prj/src/main/java/com/example/demo/CookingRecipeServiceImpl.java
@@ -12,9 +12,11 @@ public class CookingRecipeServiceImpl implements CookingRecipeService {
 	@Autowired
 	CookingRecipeDAO cookingRecipeDAO;
 	
-	public int insertRecipe(CookingRecipeDTO cookingRecipeDTO) {		
+	public int insertRecipe(CookingRecipeDTO cookingRecipeDTO) {
+		cookingRecipeDAO.deleteTemp_recipe_ingredient(cookingRecipeDTO);
 		cookingRecipeDAO.deleteTemp_recipe_content(cookingRecipeDTO);
 		cookingRecipeDAO.deleteTemp_recipe(cookingRecipeDTO);
+		cookingRecipeDAO.deleteRecipe_ingredient(cookingRecipeDTO);
 		cookingRecipeDAO.deleteRecipe_content(cookingRecipeDTO);
 		cookingRecipeDAO.deleteRecipe_img(cookingRecipeDTO);
 		cookingRecipeDAO.deleteRecipe(cookingRecipeDTO);
@@ -22,19 +24,21 @@ public class CookingRecipeServiceImpl implements CookingRecipeService {
 		int cnt=0;
 		if(cookingRecipeDAO.insertRecipe(cookingRecipeDTO)>0) {
 			if(cookingRecipeDAO.insertRecipe_content(cookingRecipeDTO)>0) {
-				cnt=1;
-				//DTO에서 이미지를 꺼내서 이름을 저장, 확장자 체크, 이미지를 Base64로 변환 및 저장
-				Util.file_nameInput(cookingRecipeDTO);
-				MultipartFile img = cookingRecipeDTO.getFoodImg();
-				if(img != null && !img.isEmpty()) {
-					if(Util.extensionCheck(cookingRecipeDTO.getFoodImgName())==-18) {
-						throw new RuntimeException("-18"); 
+				if(cookingRecipeDAO.insertRecipe_ingredient(cookingRecipeDTO)>0) {
+					cnt=1;
+					//DTO에서 이미지를 꺼내서 이름을 저장, 확장자 체크, 이미지를 Base64로 변환 및 저장
+					Util.file_nameInput(cookingRecipeDTO);
+					MultipartFile img = cookingRecipeDTO.getFoodImg();
+					if(img != null && !img.isEmpty()) {
+						if(Util.extensionCheck(cookingRecipeDTO.getFoodImgName())==-18) {
+							throw new RuntimeException("-18"); 
+						}
+
+						cnt = Util.convertImgToBase64(cookingRecipeDTO);
+						if(cnt== -20) { throw new RuntimeException("-20"); }
+
+						cnt = cookingRecipeDAO.insertImg(cookingRecipeDTO);
 					}
-					
-					cnt = Util.convertImgToBase64(cookingRecipeDTO);
-					if(cnt== -20) { throw new RuntimeException("-20"); }
-					
-					cnt = cookingRecipeDAO.insertImg(cookingRecipeDTO);
 				}
 			}
 		}
@@ -64,10 +68,28 @@ public class CookingRecipeServiceImpl implements CookingRecipeService {
 		int cnt=0;
 		if(cookingRecipeDAO.deleteRecipe_content(cookingRecipeDTO)>0) {
 			if(cookingRecipeDAO.deleteRecipe_img(cookingRecipeDTO)>0){
-				cnt=cookingRecipeDAO.deleteRecipe(cookingRecipeDTO);
+				if(cookingRecipeDAO.deleteRecipe_ingredient(cookingRecipeDTO)>0) {
+					cnt=cookingRecipeDAO.deleteRecipe(cookingRecipeDTO);
+				}
 			}
 		}
 		if(cnt==0) { throw new RuntimeException("0"); }
 		return cnt;
+	}
+	
+	public int updatePost(CookingRecipeDTO cookingRecipeDTO) {
+		cookingRecipeDAO.deleteRecipe_ingredient(cookingRecipeDTO);
+		cookingRecipeDAO.updateTitle(cookingRecipeDTO);
+		cookingRecipeDAO.updateContent(cookingRecipeDTO);
+		if(cookingRecipeDAO.insertRecipe_ingredient(cookingRecipeDTO)<1) {
+			throw new RuntimeException("0");
+		}
+		
+		MultipartFile img = cookingRecipeDTO.getFoodImg();
+		if(img != null && !img.isEmpty()) {
+			cookingRecipeDAO.updateImg(cookingRecipeDTO);
+		}
+		
+		return 1;
 	}
 }
